:judge: :memo: LICENSE: License OoI technical assessment under AGPL-3.0
This commit adds the GNU Affero General Public License Version 3 as the
license for the assessment. This is in line with the OoI's approach to
Open Source innovation¹ and also aligns with the Digital Public Goods
Standard².

I chose the AGPL-3.0 because it is a strict copyleft license for
software. This is a technical assessment, and it is rather unique that
we publish it in the open on GitHub. If others reused this assessment
for their own purposes, I believe we want to encourage public changes in
the spirit of openness and transparency. There is intentionally no real
world use case with this repository; it is a technical assessment. The
spirit of a strict copyleft license like the AGPL-3.0 seems like an
effective license choice of capturing our intent and values with how
this technical assessment is used.

¹ — https://www.unicef.org/innovation/stories/unicefs-open-source-approach-innovation
² — https://digitalpublicgoods.net/standard/

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>